% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{build_boxplot}
\alias{build_boxplot}
\title{Build GTEx boxplot}
\usage{
build_boxplot(data, title = "Gene Expression")
}
\arguments{
\item{data}{data.frame containing columns: SMTS, SMTSD, and tpm}
}
\description{
Build GTEx boxplot
}
\examples{
\dontrun{
tdb_genes <- open_gtex_array()
tbl_results <- tdb_genes["ENSG00000202059.1",]
tbl_results <- merge(tbl_results, tbl_samples, by = "sample")
build_boxplot(tbl_results)
}
}
