% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{open_gtex_array}
\alias{open_gtex_array}
\title{Open TileDB GTEx Array}
\usage{
open_gtex_array(
 
    uri = "s3://genomic-datasets/biological-databases/data/tables/gtex-analysis-rnaseqc-gene-tpm",
  attrs = "tpm",
  verbose = FALSE
)
}
\arguments{
\item{uri}{Location of TileDB array to open}

\item{attrs}{Array attributes to include}

\item{verbose}{Logical, print message with array's location}
}
\value{

}
\description{
Open TileDB GTEx Array
}
