% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{TileDBClient}
\alias{TileDBClient}
\title{TileDB Cloud API Client}
\description{
Minimal client to submit generic UDFs.
}
\section{Authentication}{
 requires an authentication token to
communicate with TileDB Cloud.
}

\examples{
\dontrun{
cli <- TileDBClient$new()
cli

# retrieve user profile
cli$user()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url}}{TileDB Cloud API URL}

\item{\code{token}}{Access token}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TileDBClient$new()}}
\item \href{#method-print}{\code{TileDBClient$print()}}
\item \href{#method-user}{\code{TileDBClient$user()}}
\item \href{#method-udf_info}{\code{TileDBClient$udf_info()}}
\item \href{#method-submit_udf}{\code{TileDBClient$submit_udf()}}
\item \href{#method-clone}{\code{TileDBClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TileDBClient}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBClient$new(token = NULL, url = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{TileDB REST API token. This is optional if the environment
variable \code{TILEDB_REST_TOKEN} is defined, but the \code{token} argument takes
precedence if both are defined.}

\item{\code{url}}{Override the default TileDB Cloud API URL}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{TileDBClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBClient$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}

\item{\code{x}}{self}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user"></a>}}
\if{latex}{\out{\hypertarget{method-user}{}}}
\subsection{Method \code{user()}}{
Retrieve TileDB Cloud user profile
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBClient$user()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-udf_info"></a>}}
\if{latex}{\out{\hypertarget{method-udf_info}{}}}
\subsection{Method \code{udf_info()}}{
Retrieve information for a registered UDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBClient$udf_info(namespace, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{Namespace of a user or organization}

\item{\code{name}}{Registered UDF name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submit_udf"></a>}}
\if{latex}{\out{\hypertarget{method-submit_udf}{}}}
\subsection{Method \code{submit_udf()}}{
Submit a generic UDF
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBClient$submit_udf(namespace, name, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{namespace}}{Namespace of a user or organization}

\item{\code{name}}{Registered UDF name}

\item{\code{args}}{List of arguments passed to the UDF}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
